cleanup: remove flag gate for batch partition split
