+++ .github/workflows/e2e.yml
@@ -187,94 +187,6 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
           flags: go-e2e,split-${{ matrix.index }},kq-${{ matrix.experimentalKeyQueues }},db-${{ matrix.database }}
 
-  batch:
-    name: "Batch / OS: (${{ matrix.os }}), key-queues: ${{ matrix.experimentalKeyQueues }}, db: ${{ matrix.database }}, split-batch-partition: ${{ matrix.splitBatchPartition }}"
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest]
-        experimentalKeyQueues: [false, true]
-        database: [sqlite, postgres]
-        splitBatchPartition: [false, true]
-    runs-on: ${{ matrix.os }}
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: true
-
-      - name: Set up Go
-        uses: actions/setup-go@v5
-        with:
-          go-version-file: ./go.mod
-
-      - name: Start postgres container
-        if: matrix.database == 'postgres'
-        run: |
-          docker run -d --name inngest-test-postgres \
-            -e POSTGRES_USER=postgres \
-            -e POSTGRES_PASSWORD=postgres \
-            -e POSTGRES_DB=inngest_test \
-            -p 5432:5432 \
-            postgres:17
-          # Wait for postgres to be ready
-          for i in $(seq 1 30); do
-            if docker exec inngest-test-postgres pg_isready -U postgres > /dev/null 2>&1; then
-              echo "Postgres is ready"
-              break
-            fi
-            if [ "$i" = "30" ]; then
-              echo "ERROR: Postgres failed to start"
-              exit 1
-            fi
-            sleep 1
-          done
-
-      - name: Build dev server
-        run: go build -cover -o ./inngest-bin ./cmd
-
-      - name: Run batch E2E tests
-        run: |
-          echo "Running dev server"
-          mkdir $GOCOVERDIR
-          if [ "${{ matrix.database }}" = "postgres" ]; then
-            nohup ./inngest-bin dev --no-discovery --no-poll --postgres-uri "postgres://postgres:postgres@localhost:5432/inngest_test?sslmode=disable" 2> /tmp/dev-output.txt &
-          else
-            nohup ./inngest-bin dev --no-discovery --no-poll 2> /tmp/dev-output.txt &
-          fi
-          echo "Ran dev server"
-          sleep 5
-          curl http://127.0.0.1:8288/dev > /dev/null 2> /dev/null
-
-          go test ./tests/batch/... -v -count=1
-        env:
-          API_URL: http://127.0.0.1:8288
-          INNGEST_EVENT_KEY: test
-          INNGEST_SIGNING_KEY: 7468697320697320612074657374206b6579
-          INNGEST_DEV: http://127.0.0.1:8288
-          TEST_MODE: true
-          TEST_DATABASE: ${{ matrix.database }}
-          EXPERIMENTAL_KEY_QUEUES_ENABLE: "${{ matrix.experimentalKeyQueues }}"
-          EXPERIMENTAL_SPLIT_BATCH_PARTITION_BY_FUNCTION: "${{ matrix.splitBatchPartition }}"
-
-      - name: Convert coverage for tooling
-        run: |
-          ls -la $GOCOVERDIR
-          go tool covdata percent -i=$GOCOVERDIR | tee coverage.txt
-
-      - name: Upload devserver logs
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: devserver-logs-batch-kq-${{ matrix.experimentalKeyQueues }}-db-${{ matrix.database }}-split-${{ matrix.splitBatchPartition }}
-          path: /tmp/dev-output.txt
-          retention-days: 7
-
-      - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v5
-        with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          flags: batch-e2e,kq-${{ matrix.experimentalKeyQueues }},db-${{ matrix.database }},split-${{ matrix.splitBatchPartition }}
-
   api:
     name: "API / OS: (${{ matrix.os }}), key-queues: ${{ matrix.experimentalKeyQueues }}, db: ${{ matrix.database }}"
     strategy:
