+++ pkg/execution/batch/redis.go
@@ -76,18 +76,6 @@ func WithoutBuffer() RedisBatchManagerOpt {
 	}
 }
 
-// WithSplitBatchPartitionByFunction registers a gate that controls whether
-// scheduled batch jobs are enqueued to a function-scoped system partition
-// (queue.KindScheduleBatch:<functionID>) instead of the single shared
-// schedule-batch partition. When the gate returns true for an accountID, that
-// account's functions each get their own batch schedule partition, avoiding
-// cross-function head-of-line blocking on the shared system queue.
-func WithSplitBatchPartitionByFunction(fn func(ctx context.Context, accountID uuid.UUID) (enable bool)) RedisBatchManagerOpt {
-	return func(m *redisBatchManager) {
-		m.splitBatchPartitionByFunction = fn
-	}
-}
-
 // NewRedisBatchManager creates a new redis batch manager, using Redis as the backing manager.
 //
 // Note that this buffers in-memory using the defaults via [DefaultMaxBufferDuration] and
@@ -126,12 +114,6 @@ type redisBatchManager struct {
 	// When nil, appends go directly to Redis. When set, appends are buffered
 	// and flushed periodically or when the buffer is full.
 	buffer *appendBuffer
-
-	// splitBatchPartitionByFunction, when non-nil and returning true for a
-	// given accountID, causes ScheduleExecution to enqueue the batch-scheduling
-	// job to a function-scoped system partition rather than the shared
-	// schedule-batch partition.
-	splitBatchPartitionByFunction func(ctx context.Context, accountID uuid.UUID) (enable bool)
 }
 
 func (b *redisBatchManager) batchKey(ctx context.Context, evt event.Event, fn inngest.Function) (string, error) {
@@ -323,10 +305,7 @@ func (b *redisBatchManager) ScheduleExecution(ctx context.Context, opts Schedule
 	jobID := opts.JobID()
 	maxAttempts := consts.MaxRetries + 1
 
-	queueName := queue.KindScheduleBatch
-	if b.splitBatchPartitionByFunction != nil && b.splitBatchPartitionByFunction(ctx, opts.AccountID) {
-		queueName = fmt.Sprintf("%s:%s", queue.KindScheduleBatch, opts.FunctionID)
-	}
+	queueName := fmt.Sprintf("%s:%s", queue.KindScheduleBatch, opts.FunctionID)
 	err := b.q.Enqueue(ctx, queue.Item{
 		JobID:       &jobID,
 		GroupID:     uuid.New().String(),
